// frontend/src/domain/models/house-details-model.ts
export type HouseDetailsModel = {
    id: number
    homeImage: string  
    type: string
    yearBuilt: number
    heating: string
    parking: string
    lot: string
    stories: number
    price: string
    beds: number
    baths: number
    neighborhood: string
    address: string
    city: string
    zipcode: string
    available: number
    anualTax: number
    hasGarage: number
    pool: number
    virtualTourLink: string
    parcelNumber: number
    lastSold: string
    total: number
    lastVisited: string
}